Add Swup progress plugin to transition between pages
