+++ assets/scss/components/_swup-progress-bar.scss
@@ -1,0 +1,4 @@
+.swup-progress-bar {
+  height: 4px;
+  background-color: $color-tertiary;
+}
