+++ assets/scss/style.scss
@@ -100,6 +100,7 @@ html.no-js [data-aos] {
 @import "components/_titles";
 @import "components/_values";
 @import "components/_social-post";
+@import "components/_swup-progress-bar";
 
 // generic
 @import "generic/_a";
